[admin] Create requirements, .gitignore file
Added some administrative files to ensure that we're all working with
the same versions of libraries, and added a .gitignore that covers most
built / compiled file types that we don't want to be committing into the
repository.

In addition, formatted the code a bit, and added a brief overview of how
to get set up for development.